api/pyproject.toml
[project]
name = "pem-api"
version = "0.1.0"
description = "Pem API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "arq>=0.27.0",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.2",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "redis>=5.3.1",
    "sentry-sdk>=2.56.0",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.37",
    "structlog>=25.5.0",
    "svix>=1.89.0",
]
